+++ src/Blocks/CtaBlock.php
@@ -14,6 +14,11 @@ class CtaBlock extends BaseBlock
         return 'cta';
     }
 
+    public static function supports(): array
+    {
+        return ['background', 'padding'];
+    }
+
     public static function label(): string
     {
         return 'Call to action';